+++ PayPal/src/test/java/com/braintreepayments/api/paypal/PayPalInternalClientUnitTest.java
@@ -92,7 +92,7 @@ public class PayPalInternalClientUnitTest {
         PayPalVaultRequest payPalRequest = new PayPalVaultRequest(true);
         payPalRequest.setBillingAgreementDescription("Billing Agreement Description");
         payPalRequest.setMerchantAccountId("sample-merchant-account-id");
-        payPalRequest.setLandingPageType(PayPalRequest.PayPalLandingPageType.LANDING_PAGE_TYPE_BILLING);
+        payPalRequest.setLandingPageType(PayPalLandingPageType.LANDING_PAGE_TYPE_BILLING);
         payPalRequest.setDisplayName("sample-display-name");
         payPalRequest.setLocaleCode("US");
         payPalRequest.setShippingAddressRequired(true);
@@ -166,7 +166,7 @@ public class PayPalInternalClientUnitTest {
         payPalRequest.setShouldRequestBillingAgreement(true);
         payPalRequest.setBillingAgreementDescription("Billing Agreement Description");
         payPalRequest.setMerchantAccountId("sample-merchant-account-id");
-        payPalRequest.setLandingPageType(PayPalRequest.PayPalLandingPageType.LANDING_PAGE_TYPE_LOGIN);
+        payPalRequest.setLandingPageType(PayPalLandingPageType.LANDING_PAGE_TYPE_LOGIN);
         payPalRequest.setDisplayName("sample-display-name");
         payPalRequest.setLocaleCode("US");
         payPalRequest.setShippingAddressRequired(true);
@@ -567,7 +567,7 @@ public class PayPalInternalClientUnitTest {
         PayPalCheckoutRequest payPalRequest = new PayPalCheckoutRequest("1.00", true);
         payPalRequest.setIntent(PayPalPaymentIntent.AUTHORIZE);
         payPalRequest.setMerchantAccountId("sample-merchant-account-id");
-        payPalRequest.setUserAction(PayPalCheckoutRequest.PayPalPaymentUserAction.USER_ACTION_COMMIT);
+        payPalRequest.setUserAction(PayPalPaymentUserAction.USER_ACTION_COMMIT);
         payPalRequest.setRiskCorrelationId("sample-client-metadata-id");
 
         sut.sendRequest(context, payPalRequest, payPalInternalClientCallback);
